# Optimize baseline and special case performance for kSmallest