@lemire 2023-03-02
@Barbarrosa This is very good. Please consider my comments.