@Barbarrosa 2023-03-11
@staltz I identified the issue, it's an overflow when requesting larger numbers of records. I'll open a PR to fix it.