@Barbarrosa 2023-03-11
@staltz @lemire I opened https://github.com/lemire/FastPriorityQueue.js/pull/38 to address the issue.